Complete removal of login redux reducers and actions
