+++ app/javascript/src/containers/TeamProfile.jsx
@@ -42,7 +42,6 @@ const mapStateToProps = state => {
   return {
     user: state.user,
     team: state.team,
-    login: state.login,
   };
 };
