+++ app/javascript/src/containers/UserProfile.jsx
@@ -54,7 +54,6 @@ class UserProfile extends Component {
 const mapStateToProps = state => {
   return {
     user: state.user,
-    login: state.login,
   };
 };
