+++ app/javascript/src/reducers/index.js
@@ -1,5 +1,4 @@
 export { default as user } from './user';
-export { default as login } from './login';
 export { default as shifts } from './shifts';
 export { default as team } from './team';
 export { default as posts } from './posts';
